requirements.txt
requests==2.32.5
Flask==3.1.2
httpx==0.28.1
opencv-python-headless==4.12.0.88
numpy<2.3
pillow==11.3.0